pyproject.toml
[project]
name = "eduverse-ai-main-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==3.2.2",
    "email-validator>=2.3.0",
    "fastapi>=0.121.1",
    "motor>=3.7.1",
    "passlib[bcrypt]==1.7.4",
    "pydantic>=2.12.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
    "stripe>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
